Stopped git-ticket-comment From Nuking My Files Without Message Ids.
I was getting a sed syntax error, so the files weren't being saved.
Really there should be better error checking here.
Also, `sed -i` could be used.
Anyway.

I think the problem was that the '\' before the newline was escaping the newline that sed required.
It works with \\\\, though, which makes me thing it gets escaped twice somewhere.

Whatever.