Changed `grep -vil` to `grep -iL`.
So, -v doesn't quite work with -l as one might think.
Well, technically it works exactly as is logical, but still somewhat unexpected.

It picks all files that have something that doesn't match a pattern.
We want all files that have nothing that matches a pattern.

This is what -L does.